requirements.txt
alembic==1.13.3
attrs==24.2.0
blinker==1.8.2
click==8.1.7
flasgger==0.9.7.1
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.5
MarkupSafe==3.0.2
marshmallow==3.23.0
marshmallow-sqlalchemy==1.1.0
mistune==3.0.2
packaging==24.1
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.3.3
pytest-flask==1.3.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.4
wheel==0.44.0
